Day 1: Solve day 1 :3
Question why, exactly, we have training that teaches us to decode the passwords lying around everywhere, rather than telling us the passwords. The password can be changed, the method of cracking it can not really.

Check in on crabs at aquarium (they're still waiting on snailfish to clear the way)