Day 4: Solve day 4 :3
Do not attempt to understand how the following paper roll is completely unremovable:
...
.@@
@@@

It's magic forklifts, that's how.

Leave as a challenge to the reader to optimise challenge further. Almost certainly, it would be faster to represent each roll of paper with the number of rolls around it, maintaining a separate list of rolls that can be removed immediately. Each step, remove any rolls that can be, then decrement the counts of its neighbours. Add neighbours that then have only 3 neighbours to the list to remove. Hop to it, reader!

Check banana stock index. What do you mean, 0.000001 cents?!